Rework parts of the crop/pad code

+++ cradle/src/itp.rs
@@ -150,7 +150,7 @@ pub fn read(f: &[u8]) -> Result<Itp, read::Error> {
 	read::read(&mut Reader::new(f))
 }
 
-pub fn read_size(f: &[u8]) -> Result<(u32, u32), read::Error> {
+pub fn read_size(f: &[u8]) -> Result<(usize, usize), read::Error> {
 	read::read_size(&mut Reader::new(f))
 }
